commit 2021/09/19

+++ src/main/java/ru/shvets/blog/repositories/CaptchaRepository.java
@@ -12,5 +12,5 @@ public interface CaptchaRepository extends JpaRepository<CaptchaCode, Long> {
     @Modifying
     @Transactional
     @Query(nativeQuery = true, value = "delete from captcha_codes where time_to_sec(timediff(now(), time)) > ?1")
-    void deleteAllTimeGreaterThanHour(int timePeriod);
+    void deleteAllTimeGreaterThanHour(long timePeriod);
 }
